refactor: import関連のshared_examplesを作成し共通化
